+++ src/main/java/org/springframework/samples/petclinic/Profiles.java
@@ -1,0 +1,6 @@
+package org.springframework.samples.petclinic;
+
+public class Profiles {
+    public static final String EXTERNALIZED = "externalized";
+    public static final String LAUREL_PROPERTIES = "laurel-properties";
+}